requirements.txt
# Core Flask Framework
Flask==2.3.3
Flask-Mail==0.9.1
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
Flask-RESTful==0.3.10

# Authentication & Security
PyJWT==2.8.0
Werkzeug==2.3.7

# Firebase & Database Dependencies
firebase-admin==6.2.0
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0

# Communication & Real-time
python-socketio==5.9.0
python-engineio==4.7.1

# Data Processing & Analytics
pandas==2.0.3
qrcode==7.4.2
Pillow==10.0.0

# File Handling & Validation
python-magic==0.4.27
email-validator==2.0.0
python-phonenumbers==8.13.11

# Development & Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
requests==2.31.0

# Logging and Monitoring
structlog==23.1.0